requirements.txt
wolframalpha~=5.1.3
tabulate
google-adk==1.2.1
protobuf~=5.29.3
python-dotenv~=1.1.0
pytest~=8.3.4
streamlit~=1.44.1
typing_extensions~=4.12.2